feat: backend done

+++ backend/src/services/laptop-service.js
@@ -1,14 +1,19 @@
 import mongoose from "mongoose"
 import Laptop from "../model/laptop-model.js"
-import { validateLaptopData } from "../utils/validation.js"
+
 
 export const createLaptop = async(dataSet)=>{
 
-    const isValid = validateLaptopData(dataSet)
-    if (!isValid){throw new Error("bad data")}
     const {model_name, price, specifications, stock_quantity, brand_id} = dataSet
     
+    if (!mongoose.Types.ObjectId.isValid(dataSet.brand_id)){
+        throw new Error("invalid id")
+    }
 
+    const exists = await Brand.exists({ _id: brand_id })
+    if (!exists) {
+        return res.status(404).json({ message: "brand not found" })
+    }
     const newLaptop = new Laptop({model_name, price, specifications, stock_quantity, brand_id})
 
     await newLaptop.save()
@@ -19,6 +24,9 @@ export const createLaptop = async(dataSet)=>{
 export const removeLaptop = async(dataSet) =>{
 
     const id = dataSet
+    if (!mongoose.Types.ObjectId.isValid(id)){
+        throw new Error("invalid id")
+    }
     const deletedLaptop = await Laptop.findByIdAndDelete(id);
     if (!deletedLaptop){
         throw new Error("Laptop not found")
@@ -34,8 +42,18 @@ export const listLaptops = async (dataSet)=>{
 
 }
 
+export const uniqueComponents = async ()=>{
+
+    const [cpus, gpus] = await Promise.all([Laptop.distinct('specifications.cpu'),Laptop.distinct('specifications.gpu')]);
+
+    return { cpus, gpus };
+  }
+
 export const getOneLaptop = async (dataSet)=>{
-    const id = dataSet
+    const id = dataSet.laptop_id
+    if (!mongoose.Types.ObjectId.isValid(id)){
+        throw new Error("invalid id")
+    }
     const laptop = await Laptop.findById(id)
     if (!laptop){
         throw new Error("Laptop not found")
@@ -44,11 +62,10 @@ export const getOneLaptop = async (dataSet)=>{
 }
 
 export const filterLaptops = async (dataSet)=>{
-
+    //validation ? 
     const {brands, gpu, cpu, fromRam, toRam} = dataSet
-    console.log(brands)
-    const filter = {}
 
+    const filter = {}
     if(brands?.length && Array.isArray(brands)){
         filter.brand_id = {$in: brands.map(id => new mongoose.Types.ObjectId(id))
         }
